Fully working VGA Char

+++ main.c
@@ -1,11 +1,11 @@
 #include "main.h"
-uint16_t* vgaBuffA;
-uint16_t* vgaBuffB;
-uint16_t* vgaBuffNext;	/*for tim2 interupt*/
+uint8_t* vgaBuffA;
+uint8_t* vgaBuffB;
+uint8_t* vgaBuffNext;	/*for tim2 interupt*/
 int main(){
 	/*Tables initialisation*/
-	uint16_t buffA[BUFFER_SIZE];
-	uint16_t buffB[BUFFER_SIZE];
+	uint8_t buffA[BUFFER_SIZE];
+	uint8_t buffB[BUFFER_SIZE];
 	vgaBuffA = buffA;
 	vgaBuffB = buffB;
 	initTable(buffA,BUFFER_SIZE,0);
